Fix overlap between menu bar and Plash
The workaround in `Utilities.swift` was incorrect. I tried to not make it show through when you have a maximized window, but that just caused it to show 1 point of the original wallpaper instead. We just have to live with macOS having a 1 point gap between the menu bar and a maximized window.

Fixes #107